ci/feature-tests/requirements.txt
-e git+https://github.com/behave/behave@bfda54448906d919ffd3fa4424b6d3c155a5388b#egg=behave
jelastic-client==1.1.30
python-gitlab==2.10.1